modif type pour la prop ID  personne

+++ app_school/src/main/java/com/intiformation/appschool/modeles/Personnes.java
@@ -30,7 +30,7 @@ public class Personnes implements Serializable{
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="idPersonne")
-	int idPersonne;
+	Long idPersonne;
 	
 	@Column(name="Identifiant")
 	String identifiant;
